qps615: driver: Load qca-ssdk module when QCA808x is detected
These changes are required since QCA808x was not able to read the
mii bus when mdio_bus_register is done and bails out with success
leading to module being loaded in the system memory. To avoid this
the following changes are made in QPS615 driver to load the QCA808x
module once the probe completes.

Change-Id: I2a4ff9591fbca9a6eecac664f9359a3b45df7b4f
Signed-off-by: Abhishek Chauhan <abchauha@codeaurora.org>